CRUDs/requirements.txt
sqlalchemy==1.4.44
fastapi==0.88.0
uvicorn==0.18.3
psycopg2-binary
kafka-python==2.0.2

